added the start of an ability system
I added an ability controller to hold and manage all abilities added to the character. Added AbilityBase with active time and cooldowns so we can extend from it for all of our abilities that we add to the ability controller. I also made sure the input for the abilities networked and put in a sample stealth ability so there is some clear foundation.